결과조회 추가

+++ src/main/java/com/sbsft/wslapi/model/NumSet.java
@@ -15,5 +15,5 @@ public class NumSet extends DrawInfo {
     private String bonus;
     private String rip;
     private String numberUnit;
-    private int totalPrize;
+    private long totalPrize;
 }
